# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UptimePulse API
  version: 0.0.1
  description: |
    API REST de UptimePulse (monitorización de uptime). Se sirve en `/docs`
    (Swagger UI) desde la propia API.

    **Autenticación.** Dos formas, ambas por `Authorization: Bearer …`:
    - un *access token* JWT de 5 min (de `/auth/register`, `/auth/login` o
      `/auth/refresh`; el refresh token viaja en una cookie httpOnly y caduca
      a los `SESSION_IDLE_TIMEOUT_MINUTES` de inactividad, 15 por defecto,
      renovándose en cada refresh);
    - una **API key** `up_…` creada por un admin en
      `/organizations/{id}/api-keys`. Actúa en nombre de la organización con
      el rol equivalente a sus scopes (`read` → readonly, `write` → editor) y
      nunca puede administrar la organización (`/me`, equipo, claves).

    **Organización activa.** Las rutas de datos actúan sobre la organización
    indicada en `X-Organization-Id`; sin cabecera, sobre la personal del
    usuario. Con API key la fija la clave y la cabecera se ignora.

    **Errores.** Siempre `{ "error": "<mensaje en español>" }`; los 400 de
    validación pueden llevar el detalle de zod en `error`. Cada respuesta
    lleva `X-Request-Id`.

    **Rate limits.** 300 peticiones/min por IP o API key en toda la API (429);
    login y registro 10/min; `/public/status`, `/invitations/{token}`
    y `/auth/username-available` 60/min.
servers:
  - url: /api
    description: Self-hosted (detrás del nginx del frontend, mismo origen)
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: JWT de acceso o API key `up_…`.
  parameters:
    organizationHeader:
      name: X-Organization-Id
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: Organización activa. Sin ella, la personal del usuario.
    id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          description: Mensaje legible, o el detalle de validación de zod.
          oneOf: [{ type: string }, { type: object }]
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string, example: gdev }
        fullName: { type: string }
        email: { type: string, format: email }
        createdAt: { type: string, format: date-time }
    AuthResponse:
      type: object
      properties:
        user: { $ref: "#/components/schemas/User" }
        accessToken: { type: string, description: JWT de 5 minutos. }
        sessionIdleMinutes:
          type: integer
          description: |
            Minutos de inactividad tras los que caduca la sesión
            (`SESSION_IDLE_TIMEOUT_MINUTES`, 15 por defecto). Es la caducidad
            del refresh token, y cada `/auth/refresh` emite uno nuevo con la
            cuenta a cero: usar la aplicación prorroga la sesión, dejar de
            usarla la cierra. La cookie no lleva `Max-Age`, así que cerrar el
            navegador también la borra.
    RegisterInput:
      type: object
      required: [username, fullName, email, password]
      properties:
        username:
          type: string
          description: 3–30 caracteres, minúsculas/números/guiones; se normaliza. Nombres reservados (admin, status, api…) rechazados.
        fullName: { type: string, maxLength: 100 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        organizationName: { type: string, description: "Nombre de la organización personal (por defecto, el del usuario)." }
    LoginInput:
      type: object
      required: [identifier, password]
      properties:
        identifier: { type: string, description: Email o nombre de usuario. }
        password: { type: string }
...
    ApiKey:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        keyPrefix: { type: string, example: up_3f9a1c2b }
        scopes: { type: array, items: { type: string, enum: [read, write] } }
        lastUsedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    ApiKeyCreated:
      allOf:
        - $ref: "#/components/schemas/ApiKey"
        - type: object
          properties:
            key:
              type: string
              description: La clave completa. Solo se devuelve aquí, una vez.
              example: up_3f9a1c2b…(64 hex)
...
  responses:
    Unauthorized:
      description: Sin token, token inválido/caducado o API key revocada.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Sin rol suficiente, organización ajena, o acción de persona intentada con API key.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: No existe o no pertenece a la organización activa.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    BadRequest:
      description: Validación.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    TooMany:
      description: Rate limit.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Registro (crea también la organización personal)
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RegisterInput" } } }
      responses:
        "201":
          description: Registrado; cookie `uptimepulse_refresh` (httpOnly, Path=/auth).
          content: { application/json: { schema: { $ref: "#/components/schemas/AuthResponse" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: "Email o username ya en uso." }
        "429": { $ref: "#/components/responses/TooMany" }
  /auth/login:
    post:
      tags: [Auth]
      summary: Login por email o username
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/LoginInput" } } }
      responses:
        "200":
          description: Sesión iniciada; cookie de refresh.
          content: { application/json: { schema: { $ref: "#/components/schemas/AuthResponse" } } }
        "401": { description: "Credenciales inválidas." }
        "429": { $ref: "#/components/responses/TooMany" }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Nuevo access token a partir de la cookie de refresh
      security: []
      responses:
        "200": { description: "Nuevo token.", content: { application/json: { schema: { $ref: "#/components/schemas/AuthResponse" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /auth/username-available:
    get:
      tags: [Auth]
      summary: ¿Está libre este username?
      security: []
      parameters:
        - { name: username, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Resultado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  username: { type: string }
                  available: { type: boolean }
                  reason: { type: string }
  /me:
    get:
      tags: [Auth]
      summary: Usuario de la sesión
      security: [{ bearer: [] }]
      responses:
        "200": { description: "Usuario.", content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    patch:
      tags: [Auth]
      summary: Cambiar username o nombre
      security: [{ bearer: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                fullName: { type: string }
      responses:
        "200": { description: "Actualizado.", content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
        "409": { description: "Username en uso." }
...
  /organizations/{id}/api-keys:
    get:
      tags: [API keys]
      summary: Listar (admin con sesión)
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/organizationHeader" }]
      responses:
        "200": { description: "Solo prefijos.", content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/ApiKey" } } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [API keys]
      summary: Crear (admin con sesión). La clave completa solo se devuelve aquí.
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/organizationHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name: { type: string, maxLength: 100 }
                scopes: { type: array, minItems: 1, items: { type: string, enum: [read, write] } }
      responses:
        "201": { description: "Creada.", content: { application/json: { schema: { $ref: "#/components/schemas/ApiKeyCreated" } } } }
        "422": { description: "Máximo 20 claves por organización." }
  /organizations/{id}/api-keys/{keyId}:
    delete:
      tags: [API keys]
      summary: Revocar (inmediato)
      security: [{ bearer: [] }]
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { name: keyId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: "Revocada." }
        "404": { $ref: "#/components/responses/NotFound" }
...